+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Store backtraces for failed background data migrations jobs
 - Make `add_column` idempotent
 - Fix `NullLockRetrier#with_lock_retries` method signature